Slightly less wrong handling of Customization
Customization.current being defined in init breaks Rake tasks when
Customization doesn't exist in the DB - tasks like db:create and
db:schema:load.

Thus, this commit moves initialization of Customization.current into
the method `Customization.current=` itself. This should not be taken as
an ultimately correct solution, merely a step taken to allow the
software to be installed and run and operate more or less correctly.